requirements.txt
absl-py==2.2.2
agents==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.44.0
anyio==4.9.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.2.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
cachetools==5.5.2
captcha==0.7.1
cattrs==24.1.3
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cover-agent @ git+https://github.com/Codium-ai/cover-agent.git@9772720361c94abe29b71af1d5e0c44d33ce02fa
coverage==7.9.1
cryptography==45.0.5
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
diff_cover==9.2.4
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
docstring-to-markdown==0.16
docstring_parser==0.16
dunamai==1.23.1
dynaconf==3.2.10
executing==2.2.0
extra-streamlit-components==0.1.80
fastjsonschema==2.21.1
feedparser==6.0.11
field==0.2.0
filelock==3.18.0
flake8==7.3.0
flatbuffers==25.2.10
fonttools==4.57.0
forex-python==1.9.2
fqdn==1.5.1
freezegun==1.5.2
frozenlist==1.6.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-aiplatform==1.89.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grep-ast==0.3.3
griffe==1.7.3
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
h5py==3.13.0
hf-xet==1.1.7
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jedi-language-server==0.41.4
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
litellm==1.67.0
loguru==0.7.3
lsprotocol==2023.0.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mcp==1.9.4
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.3
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
multiprocess==0.70.16
mypy==1.16.1
mypy_extensions==1.1.0
namex==0.0.9
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.3
notebook_shim==0.2.4
numpy==1.26.4
openai==1.91.0
openai-agents @ git+https://github.com/openai/openai-agents-python.git@18cb55e4c45f096d00263c3561d6f22a2299368e
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-stubs==2.3.0.250703
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.7
plotly==6.0.1
pluggy==1.5.0
poetry-dynamic-versioning==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.4.0
pygls==1.3.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-benchmark==5.1.0
pytest-cov==6.2.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
responses==0.25.7
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.24.0
rsa==4.9.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
s3transfer==0.11.5
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.33.0
Send2Trash==1.8.3
sentry-sdk==2.26.1
setproctitle==1.3.5
setuptools==78.1.0
sgmllib3k==1.0.0
shapely==2.1.0
simpleaudio==1.0.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soupsieve==2.6
SQLAlchemy==2.0.40
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.47.2
streamlit==1.44.0
streamlit-authenticator==0.4.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.8.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
tools==1.0.2
torc==0.1.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.0
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
trio==0.30.0
trio-websocket==0.12.2
tsne-torch==1.0.1
types-python-dateutil==2.9.0.20241206
types-pytz==2025.2.0.20250809
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
wandb==0.17.9
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
